refactor: create barrel for useHomeState
